Introduce timing information on verification + add component tests
Users attempting to profile the time it takes to verify currently
need to measure both time to generating a signature base, as well
as time to actually perform message verification using the underlying algorithm.

This change augments the return signature of `verify()` to return
benchmarks of both operations, so that drill-down views into overhead are possible.

I've also revisited some of the `component` library changes to add a
more expansive suite of tests. The tests revealed a few typos scattered
here and there that led to invalid semantics, as well as validated the
`sfv` library as correctly handling duplicate parameters.